Temp runtime error fix NOT FOR PROD
SHOULD NOT GO TO PROD. Just there so we don't lose our minds waiting for the app to close.  There's a global variable made. It's really bad. I feel bad just remembering I wrote that. It makes be feel weird inside looking at it. It's not even a meme. Just bad.